requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.1.3
Automat==0.8.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-appconf==1.0.3
django-cleanup==4.0.0
django-cropperjs==0.0.5
django-import-export==2.0.1
django-online-users==0.3
djangorestframework==3.11.0
docker==4.1.0
et-xmlfile==1.0.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
MarkupPy==1.14
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.3
pilkit==2.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
pyOpenSSL==19.1.0
pypiwin32==223
pytz==2019.3
pywin32==227
PyYAML==5.3
requests==2.22.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
websocket-client==0.57.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.7.1
